Removing over-policing of log information. Adding some logic to ensure buffer emptiness is the cause of a drop out and not a seek.
While there are definitely behavioral improvements here, the app is demonstrating lackluster results when trying to recover from a dropout in the background. Investigation into a solution for this will require some more time so in the meantimme we'll do what we can by examining analytics from our focus group and field tests.