This branch is filled with some experimental additions surrounding stream handling that I didn't want to force into the 3.1.0 release without properly testing their validity.
This commit introduces a concept of a hard fail that takes down the current audio session after 30 seconds if it hasn't recovered. After this point, user intervention is required to reboot the stream. Making decisions as to when to try to automatically perform this intervention on the user's behalf proves ineffective, as the user will know best when he or she supposedly has a connection better than we can determine in the app. The most important change in this branch is the attempt to seek back to the expected time in the stream if we see that AVPlayer has received two frames of audio data whose timestamps are not sequential (within a certain tolerance). In build 28, we are simply logging this abnormality, but build 29 attempts a seek.