Final cumulative commit for this branch before merging with master.
Added a new metric that should hopefully never be called, but it is labeled 'driftIncreasing' and signifies when AVPlayer's max seekable date drifts higher than our threshold of a 270 second behind-live-is-still-live parameter. This commit also contains this boosted parameter up from 240 seconds. Also in this commit, buffer emptiness and messages that the stream is unlikely to keep up are ignored in instances when the user is returning to the app when playback is paused. I found these callbacks in this context are illegitimate, and are artifacts of the OS attempting to deflate our app state when the app is being backgrounded in a paused state by freeing non-playing audio resources, and then re-inflating the state when the app is returned to. Thus, actions to handle a real dropout that are preceded by these callbacks should be ignored.