Work on "live" time setting
SessionManager's `vLive` returns a date that we consider to be "live" from
the point of view of the stream. That date gets set based on the max seekable
date from AVPlayer, and then is ticked up to try and avoid jumps as the
buffer changes. This work is intended to clean up the path by which that
value gets set.

* Allow forward button to be active during on-demand
* Cut down on how often we're calling `prettifyBehindLiveStatus`