WIP on going all-in on the Swift player class
Goal is to replace all AVPlayer functionality in the AudioManager with
hooks into the AudioPlayer class.

Big-picture: AudioManager becomes a place to manage AudioPlayer objects.
When events happen in an individual AudioPlayer, AudioManager can decide
what to do.

As part of this, I've temporarily ripped out a good percentage of the
recovery techniques that were in AudioManager. Once basic functionality is
solid we'll re-add that work.