WIP on detecting when stream dates move in a non-sensical way
Detect when the seekable range moves unexpectedly, though this commit
doesn't include any attempt to do anything about it. This seems to happen
at times when the network connectivity has dropped out for a period.

Eventually we should be able to use this signal to destroy our current
player and create a new one.